requirements.txt
backports.ssl-match-hostname==3.4.0.2
certifi==14.5.14
cffi==0.9.2
ConfigArgParse==0.9.3
cryptography==0.8.1
enum34==1.0.4
Flask==0.10.1
gnureadline==6.3.3
httplib2==0.9
ipython==3.1.0
itsdangerous==0.24
Jinja2==2.7.3
jsonschema==2.4.0
lxml==3.4.2
MarkupSafe==0.23
matplotlib==1.4.2
mistune==0.5.1
mitmproxy==0.11.3
mock==1.0.1
Mopidy==1.0.0
Mopidy-Spotify==1.3.0
netlib==0.11.2
nose==1.3.4
npyscreen==4.8.6
numpy==1.9.1
passlib==1.6.2
Pattern==2.6
Pillow==2.8.1
ptyprocess==0.4
pyasn1==0.1.7
pycparser==2.10
pygame==1.9.2a0
Pygments==2.0.2
Pykka==1.2.0
pymongo==3.0
pyOpenSSL==0.14
pyparsing==2.0.3
pyserial==2.7
pyspotify==1.12
python-dateutil==2.4.0
python-mpd==0.3.0
python-whois==0.3.1
pytz==2014.10
pyvbox==0.2.0
PyYAML==3.11
pyzmq==14.5.0
scikit-image==0.10.1
scipy==0.15.1
six==1.9.0
terminado==0.5
tornado==4.1
uiautomator==0.1.34
urllib3==1.10.2
urwid==1.3.0
virtualenv==12.0.6
Werkzeug==0.10.4
